Move model classes to org.nwolfhub.easycli.model
