+++ src/main/java/org/nwolfhub/easycli/Defaults.java
@@ -1,5 +1,8 @@
 package org.nwolfhub.easycli;
 
+import org.nwolfhub.easycli.model.Border;
+import org.nwolfhub.easycli.model.Template;
+
 public class Defaults {
     public static Border squareBorder = new Border("-", "-", "|");
     public static Template defaultTemplate = new Template("default", "", "\n");
