--- src/main/java/org/nwolfhub/easycli/Variable.java
@@ -1,5 +1,0 @@
-package org.nwolfhub.easycli;
-
-public class Variable {
-    
-}
