+++ src/main/java/org/nwolfhub/easycli/model/Border.java
@@ -1,4 +1,4 @@
-package org.nwolfhub.easycli;
+package org.nwolfhub.easycli.model;
 
 public class Border {
     public String upChar;
