+++ src/main/java/org/nwolfhub/easycli/model/Variable.java
@@ -1,0 +1,4 @@
+package org.nwolfhub.easycli.model;
+
+public class Variable {
+}
